+++ doc/release-notes.md
@@ -1,6 +1,6 @@
 # 3.3.4
-- version upgrade
-- 
+- spring security removed agian
+
 # 3.3.3
 - upgrade to Spring Boot 3.3.3
 - doc folder: spring cloud, github flows removed, roles-allowed updated